fixing time issues with replay and refactored small error with converting level number to an int

+++ src/nz/ac/vuw/ecs/swen225/gp20/maze/Maze.java
@@ -29,7 +29,7 @@ public class Maze {
 	public Maze (Map<?,?> boardMap) {
 		board = new Board(boardMap);
 		player = new Player(board.findPlayer(), board.getChips());
-		levelNumber = Integer.parseInt(boardMap.get("level").toString());
+		levelNumber = (int) Double.parseDouble(boardMap.get("level").toString());
 		for(Item i : board.getLoadedInv()) { //Runs 0 times for a new game, will construct a players inventory from a loaded inventory.
 			player.addToInv(i);
 		}
